flask_with_yolo_model/flask_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics>=8.3.0
opencv-python>=4.9.0
pillow>=10.2.0
google-generativeai>=0.3.2
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
